Cylinder.extend is now Cylinder.mix
Cylinder.mix now adds to the cylinder instance, while Cylinder.extend is now a generic method to extend an object (destination) with properties from other objects (sources).

Also, every module has received a review in order to keep up with the code standards. Syntax should be a bit easier to interpret. And the Utils module doesn't depend on underscore anymore.