Templates.render will render non-existing ID as HTML
Cylinder will try to use the ID of a non-existing template as the actual template to render. This will open up new possibilities as developers will be able to render anything on `render()`.

The disadvantage of this method is that templates will never be parsed beforehand, because Cylinder doesn't know how to index that template. For proper parsing, keep using `add()` to add your templates.